backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "apify-client>=2.2.1",
    "fastapi[standard]>=0.120.1",
    "firecrawl>=4.5.0",
    "google-genai>=1.46.0",
    "gotrue>=2.12.4",
    "pydantic-ai>=1.7.0",
    "pydantic-ai-slim[google,logfire,openai]>=1.7.0",
    "pydantic-evals>=1.7.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "serpapi>=0.1.5",
    "supabase>=2.22.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
